Fix crawler message parse error

+++ core/axbot-server/src/main/java/com/github/axiangcoding/axbot/server/service/entity/CrawlerResultMessage.java
@@ -6,5 +6,5 @@ import lombok.Data;
 public class CrawlerResultMessage {
     String missionId;
     String pageSource;
-    Integer timeUsage;
+    Double timeUsage;
 }
